+++ CHANGELOG.md
@@ -9,13 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Derived `ByteRepr` for `Cow<'a, str>`
+- Derived `ByteRepr` for `Cow<'a, str>`.
+- Implemented `ByteRepr` for `Box`.
+- Implemented `StaticByteRepr` for `Box`, `Cow` and arrays.
 
 ### Changed
 
 - Replaced generic `SEND`, `RECV` and `const PROTOCOL_VERSION` parameters to a single
     `CTX` that bundles all other generics in the new `MiniUdpContext` trait to
     reduce repetition of the generics and simplify adding more generics later on.
+
+### Fixed
+- Fixed `ByteRepr` derive of nested `Cow` + `Option` types like `T(Cow<'static, Option<()>>)`.
 
 ## [0.4.0] - 2026-08-13